added form to submit new forums

+++ app/src/main/java/com/example/form/models/forum.kt
@@ -1,7 +1,9 @@
 package com.example.form.models
 
+import java.math.BigDecimal
+
 data class forum (
     val titulo: String,
     val descricao: String,
-    val views: Int
+    val views: BigDecimal
 )
